Trying to fix the merges

+++ ScreenSwing/src/se/mah/k3/User.java
@@ -12,7 +12,7 @@ public class User implements Comparable<User>{
 	private double xRel=0;
 	private double yRel=0;
 	private int lives;
-	private double delay;
+	private long delay;
 	public int userHeight = 100;
 	public int userWidth = 10;
 	
@@ -74,7 +74,7 @@ public class User implements Comparable<User>{
 	public double getDelay() {
 		return delay;
 	}
-	public void setDelay(double d) {
+	public void setDelay(long d) {
 		this.delay = d;
 	}
